+++ pkg/animal/equines.go
@@ -2,11 +2,11 @@ package animal
 
 import (
 	"github.com/ironarachne/world/pkg/resource"
-	"github.com/ironarachne/world/pkg/size"
+	"github.com/ironarachne/world/pkg/species"
 )
 
-func getEquines() []Animal {
-	animals := []Animal{
+func getEquines() []species.Species {
+	animals := []species.Species{
 		{
 			Name:       "horse",
 			PluralName: "horses",
@@ -22,15 +22,10 @@ func getEquines() []Animal {
 	}
 
 	for _, a := range animals {
-		a.AnimalType = "mammal"
-		a.IsMount = true
-		a.IsPackAnimal = true
-		a.IsScavenger = false
 		a.MinHumidity = 0
 		a.MaxHumidity = 10
 		a.MinTemperature = 0
 		a.MaxTemperature = 10
-		a.Size = size.GetCategoryByName("medium")
 		a.Resources = []resource.Resource{
 			{
 				Name:         a.Name + " hide",