+++ pkg/climate/animals.go
@@ -2,60 +2,37 @@ package climate
 
 import (
 	"github.com/ironarachne/world/pkg/animal"
+	"github.com/ironarachne/world/pkg/species"
 )
 
-func (climate Climate) getAnimals() ([]animal.Animal, error) {
+func (climate Climate) getAnimals() ([]species.Species, error) {
 	animals := climate.getFilteredAnimals()
 
-	hideAnimals := animal.ByTag("hide", animals)
-	hideAnimal := animal.Random(1, hideAnimals)
+	hideAnimals := species.ByTag("hide", animals)
+	hideAnimal := species.Random(1, hideAnimals)
 
-	animals = animal.Random(climate.MaxAnimals, animals)
+	animals = species.Random(climate.MaxAnimals, animals)
 	animals = append(animals, hideAnimal...)
 
 	return animals, nil
 }
 
-func (climate Climate) getFilteredAnimals() []animal.Animal {
+func (climate Climate) getFilteredAnimals() []species.Species {
 	animals := animal.All()
-	animals = filterAnimalsForHumidity(climate.Humidity, animals)
-	animals = filterAnimalsForTemperature(climate.Temperature, animals)
+	animals = species.FilterHumidity(climate.Humidity, animals)
+	animals = species.FilterTemperature(climate.Temperature, animals)
 	animals = climate.filterAnimalsForWater(animals)
 
 	return animals
 }
 
-func filterAnimalsForHumidity(humidity int, animals []animal.Animal) []animal.Animal {
-	var filteredAnimals []animal.Animal
+func (climate Climate) filterAnimalsForWater(animals []species.Species) []species.Species {
+	var filteredAnimals []species.Species
 
 	for _, a := range animals {
-		if a.MinHumidity <= humidity && a.MaxHumidity >= humidity {
+		if a.HasTag("aquatic") && (climate.HasLakes || climate.HasOcean || climate.HasRivers || climate.HasWetlands) {
 			filteredAnimals = append(filteredAnimals, a)
-		}
-	}
-
-	return filteredAnimals
-}
-
-func filterAnimalsForTemperature(temperature int, animals []animal.Animal) []animal.Animal {
-	var filteredAnimals []animal.Animal
-
-	for _, a := range animals {
-		if a.MinTemperature <= temperature && a.MaxTemperature >= temperature {
-			filteredAnimals = append(filteredAnimals, a)
-		}
-	}
-
-	return filteredAnimals
-}
-
-func (climate Climate) filterAnimalsForWater(animals []animal.Animal) []animal.Animal {
-	var filteredAnimals []animal.Animal
-
-	for _, a := range animals {
-		if a.IsAquatic && (climate.HasLakes || climate.HasOcean || climate.HasRivers || climate.HasWetlands) {
-			filteredAnimals = append(filteredAnimals, a)
-		} else if !a.IsAquatic {
+		} else if !a.HasTag("aquatic") {
 			filteredAnimals = append(filteredAnimals, a)
 		}
 	}