+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.8.9 (2019-11-07)
+
+- Convert animals to Species struct
+- Remove Animal struct
+
 ## 0.8.8 (2019-11-06)
 
 - Split language package into conlang, language, and writing packages